# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library Api",
        "description": "Demo Library Api<br><br>Develped by <strong>Rafael Lessa de Castro Xavier</strong>",
        "contact": {
            "email": "rafaellessacastro@hotmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/book/store": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Create Book",
                "operationId": "fba15535c3a6c714cf87df22d4f6837a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "example title"
                                            },
                                            "author": {
                                                "type": "string",
                                                "example": "example author"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "example description"
                                            },
                                            "year_edition": {
                                                "type": "number",
                                                "example": "2022"
                                            },
                                            "price": {
                                                "type": "decimal",
                                                "example": "50.48"
                                            },
                                            "category_id": {
                                                "type": "number",
                                                "example": "1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "example title",
                                    "author": "example author",
                                    "description": "example description",
                                    "year_edition": "2022",
                                    "price": "50.48",
                                    "category_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "author"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "description"
                                        },
                                        "year_edition": {
                                            "type": "number",
                                            "example": 2022
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 15.85
                                        },
                                        "category_id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Exception",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Ops, Book not created."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/book/find/{id}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get Book Details",
                "operationId": "59785806a8a4f43cccc233f2d081df65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "author"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "description"
                                        },
                                        "year_edition": {
                                            "type": "number",
                                            "example": 2022
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 15.85
                                        },
                                        "category_id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/book/update/{id}": {
            "put": {
                "tags": [
                    "Book"
                ],
                "summary": "Update Book",
                "operationId": "8efe66240e63346aa97b4b772eea13d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "example title"
                                            },
                                            "author": {
                                                "type": "string",
                                                "example": "example author"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "example description"
                                            },
                                            "year_edition": {
                                                "type": "number",
                                                "example": "2022"
                                            },
                                            "price": {
                                                "type": "decimal",
                                                "example": "50.48"
                                            },
                                            "category_id": {
                                                "type": "number",
                                                "example": "1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "example title",
                                    "author": "example author",
                                    "description": "example description",
                                    "year_edition": "2022",
                                    "price": "50.48",
                                    "category_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/book/delete/{id}": {
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "Delete Book",
                "operationId": "451b09c99cc466aa6567ca0047199465",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}